refactor(library/NonTrivialString): moves out of "/customTypes" directory
