+++ src/library/NonTrivialString/ParsingError.ts
