+++ src/library/NonTrivialString/index.ts
